pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cav4-rest-python"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [{ name = "Joao A. Ferreira", email = "joao.parana@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "authlib>=1.3.0",
    "fastapi",
    "Flask",
    "Flask-SQLAlchemy",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "requests>=2.31.0",
    "starlette",
    "streamlit>=1.32.0",
]

[project.urls]
Documentation = "https://github.com/Joao A. Ferreira/cav4-rest-python#readme"
Issues = "https://github.com/Joao A. Ferreira/cav4-rest-python/issues"
Source = "https://github.com/Joao A. Ferreira/cav4-rest-python"

[tool.hatch.version]
path = "src/cav4_rest_python/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = ["mypy>=1.0.0"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/cav4_rest_python tests}"

[tool.coverage.run]
source_pkgs = ["cav4_rest_python", "tests"]
branch = true
parallel = true
omit = ["src/cav4_rest_python/__about__.py"]

[tool.coverage.paths]
cav4_rest_python = [
    "src/cav4_rest_python",
    "*/cav4-rest-python/src/cav4_rest_python",
]
tests = ["tests", "*/cav4-rest-python/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
